# Add local beatmap lookup cache to improve performance of imports