- Closes https://github.com/ppy/osu/issues/8265

Importing from stable (194 beatmaps):

Offline (master): 30.2s
Online (master): 38.2s
With local cache: 30.6s

Note that these results are with very good internet (<50ms responses to lookups). Performance will be *magnitudes* better on lesser internet connections. This also reduces a large burden on the server infrastructure during imports, where previously over 500,000 request could be fired for a normal user's beatmap collection.

Update support would need to be added at some point, but I think this is good to go as an initial implementation (it's not supposed to have 100% coverage and will gracefully fallback when not available locally). May add tests covering download / lookup if required (potentially moving the `BeatmapUpdateQueue` out of nesting – left it there for simplicity as it uses a couple of methods in `ArchiveModelManager`).

Notes about the database itself:
- 11mb after bz2 (35mb extracted)
- Contains all ranked / approved / loved beatmaps
- Contains more metadata than is being used (can be further pruned if we deem that to be useless)